moved login routes from auth routes to users routes as login
