+++ routes/api/users.js
@@ -1,11 +1,11 @@
 const express = require('express')
 const router = express.Router();
 const {
-  check,
-  validationResult
+  check
 } = require('express-validator');
 
 const {
+  create,
   login
 } = require('../../controllers/users_controller')
 
@@ -15,6 +15,11 @@ router.post('/', [
   check('password', 'please Enter password 2with 6 or more characters').isLength({
     min: 8
   })
+], create);
+
+router.post('/login', [
+  check('email', 'Please enter a valid email!').isEmail().normalizeEmail(),
+  check('password', 'please Enter password with 6 or more characters').exists()
 ], login);
 
 module.exports = router;
